Replace previous scripts for the extraction of data
